Finished Storage
All components of storage region are finished, I added a simple test unit named 'main.c' for the storage components. It is pretty good in helping to understand the classes. I also finished creating the User class and all its components. The Storage class is missing an itirator for 'type', that itirates over all instances of given type, will create it later